forcing things to execute on the main thread, maybe fixes progressbar bug?
git-svn-id: https://svn.code.sf.net/p/osirix/code/branches/aless_coredata@8892 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ nitrogen/Sources/NSException+N2.mm
@@ -25,10 +25,10 @@
 		NSArray* addresses = [self callStackReturnAddresses];
 		if (addresses.count) {
 			void* backtrace_frames[addresses.count];
-			for (int i = addresses.count-1; i >= 0; --i)
+			for (NSInteger i = addresses.count-1; i >= 0; --i)
 				backtrace_frames[i] = (void *)[[addresses objectAtIndex:i] unsignedLongValue];
 			
-			char** frameStrings = backtrace_symbols(backtrace_frames, addresses.count);
+			char** frameStrings = backtrace_symbols(backtrace_frames, (int)addresses.count);
 			if (frameStrings) {
 				for (int x = 0; x < addresses.count; ++x) {
 					if (x) [stackTrace appendString:@"\r"];
